CORE-3517 Updated "View Parameters" to support TreeSelection types.
Refactored AnalysisParameter::setParameter to use JsonUtil::getString.
Refactored AnalysisParameterViewerPanel to show only the display value of each selection in a TreeSelection argument.
Updated AnalysisParameterViewerPanel to display each value in its own row, if an argument's value is an array of values.
Currently this only affects TreeSelection and Input types.
This also allows each value to be saved in its own row when saving the parameters to a file.