CORE-3756 Readded the filter field to MyAnalysesPanel.
Added a SearchField to MyAnalysesPanel for the "name" column, used to remotely filter the AnalysesGrid with GridFilters.
Added a StringFilter for the "analysis_name" column to the GridFilters, that sets its value to match the SearchField, so that the this field is included in the remote filters in order for the SearchField to filter the grid by name or App.
Updated AnalysisServiceFacade::getAnalyses to accept a FilterPagingLoadConfig that it can use to add the filters to the URL query.