CORE-3756 Updated MyAnalysesPanel to allow filter by ID.
Added an AnalysisSearchField to wrap the logic of the AnalysisPanel Name and App filters, and added an ID filter that can alternately filter by Analysis ID.
Updated MyAnalysesPanel to use the new AnalysisSearchField and to filter by ID in the updateSelection method if the Analysis is not already loaded in the grid.
Updated AnalysisContextBuilder to return all of the Analysis info for use by the AnalysesWindowConfig, to allow access to the Analysis Name in that updateSelection method for display in the filter field.