CORE-3753 Fix for URL import validation.
The form should be allowed to submit if there is at least one valid field, but the Invalid listener on the URL TextArea fields would disable the "Upload" button if only one field was invalid.
Removed the Valid and Invalid listeners from the URL TextArea fields and added a validateForm call to the key-up listener instead.