CORE-3922 Refactored EditorTreeGrid in SharePanel to EditorGrid.
Refactored the Sharing EditorTreeGrid in SharePanel into a DataSharing EditorGrid with a CheckBoxSelectionModel.
Refactored the DiskResource List in SharePanel into a FastMap for easier lookups.
Added a USER field to the DataSharing model so that users' names can still display in the grid.
Added a SelectionListener to the SelectCollaboratorsDialog in SharePanel that will add the selected collaborators to the grid with read permissions for each DiskResource.
Added a Cancel button to the DataSharingDialog.