CORE-3922 Fixed getUpdatedSharingList checks for "varies" permissions.
Updated the loadSharingData method to use a DataSharing copy for display in the grid, since it's display permission may get changed to "varies" in this method, and that value is checked in DataSharing::equals which would cause getUpdatedSharingList to always consider the permissions for that file as updated.
Also updated addCollaborator to use a DataSharing copy for display in the grid, for consistency.
Refactored updatePermissions to not attempt to update the display model in the grid (the EditorGrid will update its models automatically anyway).