CORE-3922 Added ShareBreakdownDialog for the SharePanel.
Added a "Visibility Varies" label with an "Explain" button to the SharePanel's toolbar that is only shown when a user in the grid does not have uniform permissions on all selected files.
The "Explain" button opens the ShareBreakdownDialog that shows individual shares grouped by either user or data/disk resource.
The "Explain" panel becomes hidden once all users in the SharePanel's grid have uniform permissions for all selected resources.