CORE-3922 Updated DataSharingDialog to lookup any user's info.
Added a getUserInfo endpoint call to UserSessionServiceFacade.
Added a getUserInfo method to CollaboratorsUtil that calls the new endpoint method.
Updated DataSharingDialog to load share permissions first, then load user info with the new getUserInfo method and the usernames found in the permissions list, since users can share with users other than collaborators now.
Updated the showCurrentCollborators method in SelectCollaboratorsDialog to load collaborators with CollaboratorsUtil if they are not already loaded.