CORE-4079 Added default output folder validation.
Added setPathValidator and isValid methods to DiskResourceSelector.
Updated UserSettingPanel to not allow the user to select a default analysis output folder with a space in the path using the new DiskResourceSelector methods.
If an invalid path is selected, an error message is displayed and the selected folder is reset to the saved setting.