CORE-4205 Allow Notifications to filter Analyses window.
Updated AnalysesPresenter with setSelectedAnalyses method and a go method that can accept a list of analyses to be selected, either immediately if the analyses are already loaded, or via the AnalysisSearchField.
Updated MyAnalysesWindow to set selected analyses in the presenter from an AnalysisWindowConfig.
Refactored AnalysisWindowConfig to use a list of Analysis instead of just HasId, so if the AnalysisSearchField needs to filter a specific analysis, it can set its name in the text field.
Added NotificationAnalysisContext AutoBean with just a name and ID for use by the NotificationCallback.
Added setSelectedAnalyses method to AnalysesView.
Refactored loadAnalyses in AnalysesView with no params and to use the grid's loader in AnalysesViewImpl.
Minor refactor and updated documentation in AnalysisSearchField.