IN PROGRESS - issue CORE-4126: Add system message display area to DE. https://pods.iplantcollaborative.org/jira/browse/CORE-4126
Worked out issues related to selection updating.  

When an message is deleted by the user, if it is currently selected, the
next message will be selected.  If the deleted message was not selected,
the currently selected message will remain selected.

When a message is dismissed from the store, it will remain in the
displayed list as long as it is selected.