IN PROGRESS - issue CORE-4296: The DE needs to have a window for viewing system messages. https://pods.iplantcollaborative.org/jira/browse/CORE-4296
Fixed the SystemMessagesWindow class so that it is size appropriately,
uses deferred binding to create the internal view and calls stop() on
the Presenter when the window is hidden.

The system message cache now starts syncing when the presenter's go()
method is called and stops with the stop() method is called.

The cache now retrieves only new messages after the preliminary fetch of
messages.