CORE-4374 Fixed "View Outputs" bug for Data Window.
Fixed an issue where the SelectFolderByIdLoadHandler lazy-loading logic would be confounded by the window.refresh() call in DEWindowManager, if the Data Window was already opened and some other folder (not a parent of the folder to lazy-load) was selected.
The logic to lazy-load and select the correct folder from a new window config is already handled by the Presenter, so the refresh call is here no longer necessary.
Since the Data Window is the only window overriding the base refresh method, it is now obsolete.