CORE-4166 Fixes for windows opening outside view.
Added an adjustPosition method to iPlant windows that takes a point representing a potential position for that window, and returns an adjusted position that accounts for the window's width and height.
Refactored DEWindowManager to use this new method when positioning new windows.
Updated CenterDesktopLayout to use this new method when positioning windows during layout.
Refactored CascadeDesktopLayout to use this new method when positioning windows during layout, and to use a private Point field to keep track of the next window's position, similar to DEWindowManager.